rename MappingRasterLegend to RasterLegend
