+++ projects/wave-core-new/src/lib/layers/layer-list/layer-list.component.html
@@ -228,7 +228,7 @@
                             </ng-template>
 
                             <ng-template [ngSwitchCase]="ST.RASTER">
-                                <wave-mapping-raster-legend fxFlex="grow" [layer]="layer"></wave-mapping-raster-legend>
+                                <wave-raster-legend fxFlex="grow" [layer]="layer"></wave-raster-legend>
                             </ng-template>
 
                             <button mat-icon-button (click)="toggleLegend(layer)">
