+++ projects/wave-core-new/src/lib/layers/legend/legend-raster/raster-legend.component.html
