+++ projects/wave-core-new/src/lib/wave-core.module.ts
@@ -55,13 +55,12 @@ import {OlRasterLayerComponent, OlVectorLayerComponent} from './map/map-layer.co
 import {RenameLayerComponent} from './layers/rename-layer/rename-layer.component';
 import {LegendComponent} from './layers/legend/legend.component';
 import {VectorLegendComponent} from './layers/legend/legend-vector/vector-legend.component';
-import {RasterLegendComponent} from './layers/legend/legend-raster/raster-legend.component';
 import {LayerListComponent} from './layers/layer-list/layer-list.component';
 import {PointIconComponent} from './layers/layer-icons/point-icon/point-icon.component';
 import {LineIconComponent} from './layers/layer-icons/line-icon/line-icon.component';
 import {RasterIconComponent} from './layers/layer-icons/raster-icon/raster-icon.component';
 import {PolygonIconComponent} from './layers/layer-icons/polygon-icon/polygon-icon.component';
-import {CastMeasurementToContinuousPipe, MappingRasterLegendComponent} from './layers/legend/legend-raster/mapping-raster-legend.component';
+import {CastMeasurementToContinuousPipe, RasterLegendComponent} from './layers/legend/legend-raster/raster-legend.component';
 import {SafeStylePipe} from './util/pipes/safe-style.pipe';
 import {SmallTimeInteractionComponent} from './time/small-time-interaction/small-time-interaction.component';
 import {TimeConfigComponent} from './time/time-config/time-config.component';
@@ -158,7 +157,7 @@ const WAVE_COMPONENTS = [
     LoadProjectComponent,
     LoginComponent,
     MapContainerComponent,
-    MappingRasterLegendComponent,
+    RasterLegendComponent,
     MeanRasterPixelValuesOverTimeDialogComponent,
     MultiLayerSelectionComponent,
     NavigationComponent,
